requirements.txt
bcrypt
celery
epdb
flask
Flask-Bootstrap
Flask-MySQLdb
Flask-LoginManager
flask-sqlalchemy
flask-wtf
flask_admin
flask_mongoengine
flask_pymongo
flask_restful
flask_security
flower
gunicorn
pymongo